Update fetch to pull & add license headers
