+++ pkg/utils/util.go
@@ -1,12 +1,20 @@
-/*
-Copyright © 2023 NAME HERE <EMAIL ADDRESS>
-
-*/
+// Copyright 2023 Interlynk.io
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 package utils
 
 import "math/rand"
-
 
 func RandomPick(min int, max int) int {
 	return min + rand.Intn(max-min)
